// types/enums.ts
export enum Gender {
  MALE = "male",
  FEMALE = "female",
  OTHER = "other"
}

export enum Plan {
  FREE = "free",
  PREMIUM = "premium",
  VIP = "vip"
}

export enum MaritalStatus {
  NEVER_MARRIED = "never_married",
  DIVORCED = "divorced",
  WIDOWED = "widowed",
  SEPARATED = "separated"
}

export enum Religion {
  HINDUISM = "hinduism",
  ISLAM = "islam",
  CHRISTIANITY = "christianity",
  SIKHISM = "sikhism",
  BUDDHISM = "buddhism",
  JAINISM = "jainism",
  OTHER = "other"
}

export enum FamilyType {
  NUCLEAR = "nuclear",
  JOINT = "joint",
  SINGLE_PARENT = "single_parent"
}

export enum FamilyValues {
  TRADITIONAL = "traditional",
  MODERN = "modern",
  MODERATE = "moderate"
}

export enum FamilyStatus {
  UPPER_CLASS = "upperClass",
  UPPER_MIDDLE_CLASS = "upperMiddleClass",
  MIDDLE_CLASS = "middleClass",
  LOWER_MIDDLE_CLASS = "lowerMiddleClass"
}

export enum Diet {
  VEG = "veg",
  NON_VEG = "non_veg",
  EGGETARIAN = "eggetarian",
  VEGAN = "vegan"
}

export enum Habits {
  NEVER = "never",
  OCCASIONALLY = "occasionally",
  REGULARLY = "regularly"
}

export enum Exercise {
  NEVER = "never",
  OCCASIONALLY = "occasionally",
  REGULARLY = "regularly"
}

export enum SleepSchedule {
  EARLY_BIRD = "early_bird",
  NIGHT_OWL = "night_owl",
  REGULAR = "regular"
}

export enum SocialLife {
  INTROVERT = "introvert",
  EXTROVERT = "extrovert",
  AMBIVERT = "ambivert"
}

export enum Travel {
  LOVE = "love",
  LIKE = "like",
  NEUTRAL = "neutral",
  DISLIKE = "dislike"
}

export enum Pets {
  LOVE = "love",
  LIKE = "like",
  NEUTRAL = "neutral",
  DISLIKE = "dislike"
}

export enum Education {
  HIGH_SCHOOL = "high_school",
  BACHELORS = "bachelors",
  MASTERS = "masters",
  MBA = "mba",
  PHD = "phd",
  OTHER = "other"
}

export enum Industry {
  IT = "it",
  FINANCE = "finance",
  HEALTHCARE = "healthcare",
  EDUCATION = "education",
  GOVERNMENT = "government",
  OTHER = "other"
} 